HARP-13329: Move projectTilePlaneCorners out of TileGeometryCreator.
